Changed the type of "price" to BigDecimal.

+++ pojo/src/main/java/com/fengwei/pojo/ProductGpu.java
@@ -3,6 +3,7 @@ package com.fengwei.pojo;
 import javax.persistence.Column;
 import javax.persistence.Id;
 import javax.persistence.Table;
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Table(name = "product_gpu")
@@ -43,7 +44,7 @@ public class ProductGpu {
     @Column(name = "main_image")
     private String mainImage;
 
-    private Long price;
+    private BigDecimal price;
 
     private Integer stock;
 
@@ -259,14 +260,14 @@ public class ProductGpu {
     /**
      * @return price
      */
-    public Long getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
     /**
      * @param price
      */
-    public void setPrice(Long price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
